Zsh-compat fixes (setopt MONITOR NOTIFY), and general minor fixes.
The most important update is to set MONITOR and NOTIFY in Zsh to make sure the shell SIGHUPs the coprocess on exit!

The history retrieval command (fc -ln -1) could fail if zsh history was empty. Now, send error to /dev/null on such cases.

Remove some extra spacing in tip / warning messages.

Remove the use of command env, as that is unnecessary.