CTRL-C cancels current request. Fix bugs related to new python tiktoken coproc hack.
CTRL-C now only cancels current request (in case the request takes to long or hang). Behaviour after CTRL-C is pressed is still undefined, but may improve over time a little.

The tiktoken coproc would interfere with voice chat mode (option -w with -cc). Add the recorder pid needed for the wait builtin to work correctly. Also, must be sure to kill all forked children.

Setting the initial $REPLY_OLD var to the last history command was giving me some problems with zsh history, so we add a limit to how long that command may be (10000 chars).

In any case, I observed that both bash and zsh history file may get corrupted, in which case, these need to be removed (or moved for backup), and a new history file should be created from the start.