Added environmental parameters to change API URL base, and endpoint.
This is just an attempt to let the user set a different endpoint, as there are many services which use the same type of request as OpenAI API endpoints. See explanation on the man page for how to set $APIURL, and $APIURLBASE as they differ a little in how the endpoint of the API URL is accounted for.

General bash fix, when in editing mode, check that $REPLY does not contain a newline, if it does, set read with input termination as <CTRL-D>, as only the first line of multi-line input is read with bash read command, unless a CTRL-D flush is implemented.

Improve some warning messages with tips...