Fix chat cmd !help (-h) with Zsh.
Improve compatibility with Zsh when the monitor setting is on in some functions.

Also, introducing coproc python requires tweaking some functions for them to work properly... I am on it!

However, I may drop support for Zsh in coming releases, as it is too much of a pain to test all functions in both bash and zsh...