Fix some Zsh compat issues, specially python tiktoken coproc.
I was able to keep Zsh support. It seems to work, but I will drop support for Zsh for the next version release.

Trying to keep my pains low... It was a good case study, though! Just not worth maintaining Zsh compat with such a complex bash/ksh code anymore!