Bracketed paste on (paste from clipboard without multiline options -uU).
Bash readline can be set with enable-bracketed-past and a little tweak of read command can make  it work to paste multiline input from the clipboard!

Multiline option -u and cat prompter -U are still available normally.

Improve compat with older Bash versions.

We reccomend running bash 5 or newer, but the script kind of works (not all features as intended) with bash v4.1+.