Improve selecting specific files in chat session commands.
For example, the chat command "/copy . test" would copy the current history file to the glob of "test". Is there only a file created with the name "test22.tsv" in the cache directory, the glob "test" would match and select "test22.tsv".

Now, if user input is "/copy . test.tsv", and if there is not a "test.tsv" file, it will then create a new one.

 That is what I expect from this command syntax and behaviour.