Expose TTS audio replay mechanism to user.
It was shown only when there was a INT or error when playing TTS audio, but I figured that replaying the received audio is difficult to do manually. So in the name of workflow, improved and exposed the replay functionality.

Change some multiple spaces to TAB, to keep formatting more homogenous througout the script.

Also, set a variable to test for the weird mp3.. globbing in one function.