Improve user interface at the readline level.
Also, deactivate the automatic one-shot ctr-d prompter when detecting multiine input and ctd-d prompter was not active. Bracketed-paste in readline can "probably" deal with that sort of input!

Improve the /grep command and functionality to grep for the regex in all files when not found at the current history file.