Small optimisations and fix the "regenerate" command with multiline input.
Incredibly, had not tried to regenerate a request of a multiline prompt so far!
The way it was, it would mess up the chatgpt.sh tsv history file with unescaped input from the old reply !

Improve the fallback functions _escapef and _unescapef. They should both me faster now, and run under OpenBSD without the hack I was using for sed....