Re-enable option -e (--edit) to edit input on stdin, or file read.
With option -e, edit the first user prompt always. So if input is from stdin, or is read from a file, or is set as positional arguments on script invocation, edit it before the first request.

So that:

    echo "Hello, there!" | chatgpt.sh -cce       #stdin

    chatgpt.sh -cce hello.txt                    #file read

    chatgpt.sh -cce "Hello, there!"              #positional arguments

will edit the input text before the first request.

Disable __spinf fun on stdout not free to avoid excess animation of pipelines.

Disable a weyrd conditional I had written for skipping one-shot editing custom prompts on some ocasions, such as stdout not free, setting option -E (--exit), and non-multi-turn. While it was created to automate skipping custom prompt one-shot editing in these circunstances, it also represents some loss of functionality. the same behaviour can be accomplished easily by the end-user by setting "-S ..custom" instead of "-S .custom" in such instances himself. Meaning that automation in these cases are not really needed.

Also, increase a little temperature and presence penalty for the chat bot mode when these are not set by the user (in order to unlobotomise the bot). The newer models are less random and less sensitive to temperature, it seems...

Update readme.md, help, and  man pages.