Improve -HH option syntax and interface.
It was  possible to search for a session in a history file with the following syntax examples for the current (defaults) history file:

  chatgpt.sh -HH .
  chatgpt.sh -HH /session .

and for a unix.tsv hist file:

  chatgpt.sh -HH.unix
  chatgpt.sh -HH -S.unix
  chatgpt.sh -HH /unix

However, it was not possible to have it like:

  chatgpt.sh -HH.unix .
  chatgpt.sh -HH  unix

And internally, the code should not be a problem in such a situation.

So a regression which was added recently, plus some improvement was added, so  now it works.