Improve chat commands -L (log to file), -w (whisper), and -z (TTS).
Always let them ON when there is an argument, otherwise toggle them.

TTS argumenets passed through chat cmd -z were not being processed (such as changing the TTS voice, etc). Now fixed!

The logic for the whisper fun in cmd_runf() function is a bit complicated but seems to work now.

Also, I noticed that the translations endpoint (-W) (translations to English endpoint) was never being really set with chat command -W. It only worked if it was set from the command line upon script invocation, but could not be toggled.

Fix checking script options in the chat mode + whisper + TTS (the check was skipped when Whisper was set).