Improve option -e for editing first user input from pos args.
For example, if the last positional argument is a file name, load it and edit it if option -e is set.

If option -x is also set, it has always edited the first user input, so I am not going to change this behaviour.

Fix excess escaping of user input. It was a bug which would only be visible in multi line prompts and when there was an edition before request, or request error and re-editing.