This improves printing out the last session option -HH
When there is a break session mark but no content afterward. This change has to do with the code refactor of 3 months ago when we set to record to a history file only after a response. This is a leftover adjustment to that.

---
```
Update chatgpt.sh: Improve session handling by removing break marks

Developer comments: This improves printing out the last session option -HH

When there is a break session mark but no content afterward. This change has to do with the code refactor of 3 months ago when we set to record to a history file only after a response. This is a leftover adjustment to that.
```